pyproject.toml
[project]
name = "projet-4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "category-encoders>=2.9.0",
    "imblearn>=0.0",
    "joblib>=1.5.3",
    "matplotlib>=3.10.8",
    "mlflow>=1.27.0",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "psycopg2>=2.9.11",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.1",
    "seaborn>=0.13.2",
    "shap>=0.51.0",
    "tqdm>=4.67.3",
]
